+++ packages/components/src/index.ts
@@ -1,5 +1,5 @@
 import './index.css';
-import AddressForm from './systems/AddressForm';
+import PatientAddressForm from './systems/PatientAddressForm';
 import { Alert } from './particles/Alert';
 import Checkbox from './particles/Checkbox';
 import DispenseUnitSelect from './particles/DispenseUnitSelect';
@@ -85,11 +85,12 @@ import {
 } from './analytics/buildFieldSnapshot';
 import { PatientSelect } from './systems/PatientSelect';
 import { validateRealPhoneNumber } from './utils/validateRealPhoneNumber';
+import { npiRegex, zipCodeRegex } from './utils/regex';
 
 export { usePhoton, PhotonClientStore, PhotonContext };
 
 export {
-  AddressForm,
+  PatientAddressForm,
   AddressAutocompleteInput,
   Alert,
   Banner,
@@ -156,7 +157,9 @@ export {
   DRAFT_PRESCRIPTION_FORM_FIELDS,
   PatientSelect,
   isValidPrescriptionRoutingConstraint,
-  validateRealPhoneNumber
+  validateRealPhoneNumber,
+  npiRegex,
+  zipCodeRegex
 };
 
 // Export types
